More mypy fixes

+++ src/koopmans/workflows/_unfold_and_interp.py
@@ -85,7 +85,7 @@ class UnfoldAndInterpolateWorkflow(Workflow):
                     # For dscf, self.bands correspond to the supercell so band_filling involves many copies of each
                     # band
                     assert self.kpoints.grid is not None
-                    ngrid = np.prod(self.kpoints.grid)
+                    ngrid = np.prod(self.kpoints.grid, dtype=int)
                 else:
                     # For dfpt, self.bands correspond to the primitive cell so band_filling is already the correct
                     # dimensions
